+++ infra/modular-examples/aws-msft-365/main.tf
@@ -3,12 +3,12 @@ terraform {
     # for the infra that will host Psoxy instances
     aws = {
       source  = "hashicorp/aws"
-      version = "~> 4.12"
+      version = "~> 4.51"
     }
 
     # for API connections to Microsoft 365
     azuread = {
-      version = "~> 2.0"
+      version = "~> 2.3"
     }
   }
 }
@@ -62,6 +62,13 @@ moved {
   to   = module.global_secrets.aws_ssm_parameter.secret["PSOXY_ENCRYPTION_KEY"]
 }
 
+module "cognito-identity-pool" {
+  source = "../../modules/aws-cognito-pool"
+
+  developer_provider_name = "azure-access"
+  name                    = "azure-ad-federation"
+}
+
 module "msft-connection" {
   for_each = module.worklytics_connector_specs.enabled_msft_365_connectors
 
@@ -74,34 +81,31 @@ module "msft-connection" {
   required_oauth2_permission_scopes = each.value.required_oauth2_permission_scopes
 }
 
-module "msft-connection-auth" {
-  for_each = module.worklytics_connector_specs.enabled_msft_365_connectors
+module "cognito-identity" {
+  source = "../../modules/aws-cognito-identity-cli"
 
-  source = "../../modules/azuread-local-cert"
-  # source = "git::https://github.com/worklytics/psoxy//infra/modules/azuread-local-cert?ref=v0.4.10"
-
-  application_object_id = module.msft-connection[each.key].connector.id
-  rotation_days         = 60
-  cert_expiration_days  = 180
-  certificate_subject   = var.certificate_subject
+  identity_pool_id = module.cognito-identity-pool.pool_id
+  aws_region       = var.aws_region
+  login-ids        = { for k in keys(module.msft-connection) : k => "${module.cognito-identity-pool.developer_provider_name}=${module.msft-connection[k].connector.application_id}" }
+  aws_role         = var.aws_assume_role_arn
 }
 
-module "msft-365-connector-key-secrets" {
+module "msft-connection-auth-federation" {
   for_each = module.worklytics_connector_specs.enabled_msft_365_connectors
 
-  source = "../../modules/private-key-aws-parameter"
-  # source = "git::https://github.com/worklytics/psoxy//infra/modules/private-key-aws-parameter?ref=v0.4.10"
+  source = "../../modules/azuread-federated-credentials"
+  # source = "git::https://github.com/worklytics/psoxy//infra/modules/azuread-federated-credentials?ref=v0.4.11"
 
-  instance_id = each.key
-
-  private_key_id = module.msft-connection-auth[each.key].private_key_id
-  private_key    = module.msft-connection-auth[each.key].private_key
-  ssm_path       = var.aws_ssm_param_root_path
-  kms_key_id     = var.aws_ssm_key_id
+  application_object_id = module.msft-connection[each.key].connector.id
+  display_name          = "AccessFromAWS"
+  description           = "AWS federation to be used for psoxy Connector - ${each.value.display_name}${var.connector_display_name_suffix}"
+  issuer                = "https://cognito-identity.amazonaws.com"
+  audience              = module.cognito-identity-pool.pool_id
+  subject               = module.cognito-identity.identity_id[each.key]
 }
 
 # grant required permissions to connectors via Azure AD
-# (requires terraform configuration being applied by an Azure User with privelleges to do this; it
+# (requires terraform configuration being applied by an Azure User with privileges to do this; it
 #  usually requires a 'Global Administrator' for your tenant)
 module "msft_365_grants" {
   for_each = module.worklytics_connector_specs.enabled_msft_365_connectors
@@ -130,7 +134,7 @@ module "psoxy-msft-connector" {
   function_zip_hash               = module.psoxy-aws.deployment_package_hash
   api_caller_role_arn             = module.psoxy-aws.api_caller_role_arn
   aws_assume_role_arn             = var.aws_assume_role_arn
-  example_api_calls               = each.value.example_calls
+  example_api_calls               = each.value.example_api_calls
   aws_account_id                  = var.aws_account_id
   path_to_repo_root               = var.psoxy_base_dir
   todo_step                       = module.msft_365_grants[each.key].next_todo_step
@@ -146,8 +150,18 @@ module "psoxy-msft-connector" {
       CLIENT_ID            = module.msft-connection[each.key].connector.application_id
       REFRESH_ENDPOINT     = module.worklytics_connector_specs.msft_token_refresh_endpoint
       PSEUDONYMIZE_APP_IDS = tostring(var.pseudonymize_app_ids)
+      IDENTITY_POOL_ID     = module.cognito-identity-pool.pool_id,
+      IDENTITY_ID          = module.cognito-identity.identity_id[each.key]
+      DEVELOPER_NAME_ID    = module.cognito-identity-pool.developer_provider_name
     }
   )
+}
+
+resource "aws_iam_role_policy_attachment" "cognito_lambda_policy" {
+  for_each = module.worklytics_connector_specs.enabled_msft_365_connectors
+
+  role       = module.psoxy-msft-connector[each.key].instance_role_name
+  policy_arn = module.cognito-identity-pool.policy_arn
 }
 
 module "worklytics-psoxy-connection-msft-365" {
@@ -373,4 +387,8 @@ locals {
 
 output "instances" {
   value = local.all_instances
+}
+
+output "cognito-identities" {
+  value = module.cognito-identity
 }