+++ infra/modules/aws-cognito-pool/variables.tf
@@ -1,0 +1,9 @@
+variable "developer_provider_name" {
+  type        = string
+  description = "Custom provider name to use"
+}
+
+variable "name" {
+  type        = string
+  description = "Name of the pool"
+}